Refactor graphene fields and types
