+++ apps/contextualupdate/schema.py
@@ -5,7 +5,8 @@ from graphene_django_extras import PageGraphqlPagination, DjangoObjectField
 from apps.crisis.enums import CrisisTypeGrapheneEnum
 from apps.contextualupdate.models import ContextualUpdate
 from apps.contextualupdate.filters import ContextualUpdateFilter
-from utils.fields import DjangoPaginatedListObjectField, CustomDjangoListObjectType
+from utils.graphene.types import CustomDjangoListObjectType
+from utils.graphene.fields import DjangoPaginatedListObjectField
 
 
 class ContextualUpdateType(DjangoObjectType):
