+++ apps/parking_lot/schema.py
@@ -5,7 +5,8 @@ from graphene_django_extras import PageGraphqlPagination, DjangoObjectField
 from apps.parking_lot.models import ParkedItem
 from apps.parking_lot.enums import ParkedItemGrapheneEnum
 from apps.parking_lot.filters import ParkingLotFilter
-from utils.fields import DjangoPaginatedListObjectField, CustomDjangoListObjectType
+from utils.graphene.types import CustomDjangoListObjectType
+from utils.graphene.fields import DjangoPaginatedListObjectField
 
 
 class ParkedItemType(DjangoObjectType):
